PR指摘の import/export テストを補強
