+++ src/features/options/lib/import-export.restore.test.ts
@@ -1,7 +1,6 @@
-// @vitest-environment jsdom
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 
-import type { CustomProject, UserSettings } from '@/types/storage'
+import type { UserSettings } from '@/types/storage'
 
 vi.mock('@/lib/storage/categories', () => ({
   saveParentCategories: vi.fn(),
@@ -46,107 +45,11 @@ import { getUserSettings } from '@/lib/storage/settings'
 import { createOrUpdateUrlRecordsBatch } from '@/lib/storage/urls'
 
 import { importSettings } from './import-export'
-
-type StorageStore = Record<string, unknown>
-
-const clone = <T>(value: T): T => {
-  if (value === undefined) {
-    return value
-  }
-  // eslint-disable-next-line unicorn/prefer-structured-clone
-  return JSON.parse(JSON.stringify(value)) as T
-}
-
-const readStorageByKeys = (
-  store: StorageStore,
-  keys?: string | string[] | Record<string, unknown>,
-) => {
-  if (keys === undefined) {
-    return clone(store)
-  }
-
-  if (typeof keys === 'string') {
-    return { [keys]: clone(store[keys]) }
-  }
-
-  if (Array.isArray(keys)) {
-    const result: Record<string, unknown> = {}
-    for (const key of keys) {
-      result[key] = clone(store[key])
-    }
-    return result
-  }
-
-  const result: Record<string, unknown> = {}
-  for (const [key, fallback] of Object.entries(keys)) {
-    result[key] = store[key] === undefined ? clone(fallback) : clone(store[key])
-  }
-  return result
-}
-
-const createChromeMock = (initialStore: StorageStore = {}) => {
-  const store = clone(initialStore)
-  const get = vi.fn(
-    async (keys?: string | string[] | Record<string, unknown>) =>
-      readStorageByKeys(store, keys),
-  )
-  const set = vi.fn(async (next: Record<string, unknown>) => {
-    for (const [key, value] of Object.entries(next)) {
-      store[key] = clone(value)
-    }
-  })
-
-  ;(globalThis as unknown as { chrome: typeof chrome }).chrome = {
-    storage: {
-      local: { get, set },
-    },
-    i18n: {
-      getUILanguage: () => 'ja',
-    },
-    runtime: {
-      getManifest: () => ({ version: '9.9.9' }),
-    },
-  } as unknown as typeof chrome
-
-  return { store }
-}
-
-const buildFullUserSettings = (
-  override: Partial<UserSettings> = {},
-): UserSettings => ({
-  removeTabAfterOpen: true,
-  removeTabAfterExternalDrop: true,
-  excludePatterns: ['existing-pattern'],
-  enableCategories: true,
-  autoDeletePeriod: 'never',
-  showSavedTime: false,
-  clickBehavior: 'saveSameDomainTabs',
-  excludePinnedTabs: true,
-  openUrlInBackground: true,
-  openAllInNewWindow: false,
-  confirmDeleteAll: false,
-  confirmDeleteEach: false,
-  colors: {},
-  ollamaModel: '',
-  ...override,
-})
-
-const buildCustomProject = (
-  override: Partial<CustomProject> = {},
-): CustomProject => ({
-  id: 'project-1',
-  name: 'Project 1',
-  projectKeywords: {
-    titleKeywords: [],
-    urlKeywords: [],
-    domainKeywords: [],
-  },
-  urlIds: [],
-  categories: [],
-  createdAt: 1,
-  updatedAt: 1,
-  ...override,
-})
+import {
+  buildCustomProject,
+  buildFullUserSettings,
+  createChromeMock,
+} from './import-export.test-fixtures'
 
 describe('import/export restore regression', () => {
   beforeEach(() => {
@@ -206,19 +109,26 @@ describe('import/export restore regression', () => {
       ],
     })
     vi.mocked(getUserSettings).mockResolvedValue(buildFullUserSettings())
-    vi.mocked(createOrUpdateUrlRecordsBatch).mockResolvedValue(
-      new Map([
-        [
-          'https://docs.example.com/release-notes',
-          {
-            id: 'imported-url',
-            url: 'https://docs.example.com/release-notes',
-            title: 'Release Notes',
-            savedAt: 600,
-          },
-        ],
-      ]),
-    )
+    const existingUrlRecord = {
+      id: 'existing-url',
+      url: 'https://docs.example.com/start',
+      title: 'Start',
+      savedAt: 500,
+    }
+    const importedUrlRecord = {
+      id: 'imported-url',
+      url: 'https://docs.example.com/release-notes',
+      title: 'Release Notes',
+      savedAt: 600,
+    }
+    vi.mocked(createOrUpdateUrlRecordsBatch).mockImplementation(async () => {
+      await chrome.storage.local.set({
+        urls: [existingUrlRecord, importedUrlRecord],
+      })
+      return new Map([
+        ['https://docs.example.com/release-notes', importedUrlRecord],
+      ])
+    })
 
     const result = await importSettings(
       JSON.stringify({
@@ -287,6 +197,7 @@ describe('import/export restore regression', () => {
         subCategoryOrderWithUncategorized: ['Reading', 'Release'],
       },
     ])
+    expect(store.urls).toEqual([existingUrlRecord, importedUrlRecord])
     expect(store.customProjectOrder).toEqual([
       'existing-project',
       'imported-project',
